added first algoithm
I made the robots move autonomously, so when I make the algoithms, they would search on their own like they will in the competition . Also, I added the first algorithm. It's pretty basic and I put the description of how the algorithm works on a doc on our google drive called "algorithms". It is pretty basic and there are some things that I still need to clear up using this algorithm but so far it works. I  haven't tested it with any other algoithm, I just wanted a basic algorithm that worked.